pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "microsoft-opentelemetry"
version = "0.1.0a2"
description = "Minimal template for the Microsoft OpenTelemetry distro for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-core<2.0.0,>=1.28.0",
    "azure-core-tracing-opentelemetry~=1.0.0b11",
    "azure-monitor-opentelemetry-exporter~=1.0.0b49",
    "opentelemetry-sdk==1.40",
    "opentelemetry-api==1.40",
    "opentelemetry-exporter-otlp-proto-http==1.40",
    "opentelemetry-instrumentation==0.61b0",
    "opentelemetry-instrumentation-django==0.61b0",
    "opentelemetry-instrumentation-fastapi==0.61b0",
    "opentelemetry-instrumentation-flask==0.61b0",
    "opentelemetry-instrumentation-psycopg2==0.61b0",
    "opentelemetry-instrumentation-requests==0.61b0",
    "opentelemetry-instrumentation-urllib==0.61b0",
    "opentelemetry-instrumentation-urllib3==0.61b0",
    "opentelemetry-instrumentation-logging==0.61b0",
    "opentelemetry-instrumentation-openai-agents-v2==0.1.0",
    "opentelemetry-instrumentation-openai-v2==2.3b0",
    "opentelemetry-resource-detector-azure<1.0.0,>=0.1.5",
    "langchain-core>=0.2.0",
    "wrapt>=1.0",
    "opentelemetry-util-genai>=0.3b0",
    "microsoft-agents-activity>=0.9.0",
    "microsoft-agents-hosting-core>=0.9.0",
    "aiohttp>=3.8.0",
    "PyJWT",
    "requests"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
dev = [
    "black>=24.0",
    "pylint>=3.3",
    "mypy>=1.10",
    "pyright>=1.1",
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=2.0",
]

[project.entry-points.opentelemetry_instrumentor]
langchain = "microsoft.opentelemetry._genai._langchain._tracer_instrumentor:LangChainInstrumentor"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.pylint.main]
py-version = "3.10"

[tool.pylint.format]
max-line-length = 120

[tool.coverage.run]
source = ["microsoft"]

[tool.coverage.report]
show_missing = true

[tool.pylint.messages_control]
disable = [
    "missing-function-docstring",
    "missing-module-docstring",
    "missing-class-docstring",
    "import-outside-toplevel",
    "import-error",
    "no-name-in-module",
    "duplicate-code",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-many-locals",
    "wrong-import-order",
    "ungrouped-imports",
    "unknown-option-value",
    "fixme",
    "invalid-name",
    "unused-argument",
    "protected-access",
    "too-many-lines",
    "too-many-public-methods",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
